Remove useless cast that would break Windows 64-bit compilation.
This cast broke Windows 64-bit compilation with a hard error.
$ x86_64-w64-mingw32-g++ --version
x86_64-w64-mingw32-g++ (GCC) 4.8.3
Error message:
widget_panel.cxx:82:41: error: cast from ‘void*’ to ‘long int’ loses precision [-fpermissive]
 wCallback->do_callback(wCallback, (long)v);
                                         ^


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@10668 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/widget_panel.cxx
@@ -78,8 +78,8 @@ Fl_Menu_Item menu_3[] = {
 Fl_Input *v_input[4]={(Fl_Input *)0};
 
 static void cb_1(Fl_Tile*, void* v) {
-  wComment->do_callback(wComment, (long)v);
-wCallback->do_callback(wCallback, (long)v);
+  wComment->do_callback(wComment, v);
+wCallback->do_callback(wCallback, v);
 }
 
 Fl_Text_Editor *wComment=(Fl_Text_Editor *)0;
